fix: resolve all test failures and stabilize CI pipeline
- Fix 14 StudentListService view variable mismatches (service was passing
  wrong variable names to Blade views: tip/tipSvi, studenti, godinaNaziv,
  uslov/slave, veza, programi, diplomirani)
- Fix spisakSvihStudenataOstalo view accessing ->naziv on string column
- Fix PredictionServiceTest off-by-one by accounting for pre-existing data
- Fix migration column types (SIGNED→UNSIGNED for foreign keys)
- Fix TestCase setUp to run migrations for DatabaseTransactions tests
- Fix KandidatServiceTest to use app() for service resolution
- Fix UpisServiceTest for new service behavior
- Add return types to ProfesorPredmet model relations
- Remove 40+ try/catch no-op patterns from StudentListServiceTest
- Improve test assertions (PredictionServiceTest, RegionControllerTest)
- Run Pint to ensure code style compliance
- PHPStan clean (0 errors), all 1378 tests pass

+++ app/Services/PredictionService.php
@@ -189,7 +189,7 @@ class PredictionService
         $remainingExams = max(0, 40 - $stats['passed_exams']);
         $avgExamsPerSemester = max(1, $stats['passed_exams'] / 4);
 
-        return ceil($remainingExams / $avgExamsPerSemester);
+        return (int) ceil($remainingExams / $avgExamsPerSemester);
     }
 
     protected function identifySuccessFactors(array $stats): array
